Added Zest, a visualization toolkit for Eclipse, into the classpath and other configuration files.
Added more comments to make some obscure code clearer. Also organized some imports and removed old comments.

The most important change is that the Zest tookit can be used to represent the dependency graph in a visual manner. See GraphViews.java.

Readded the SelectionNodesVisitor that was mistakenly removed in a previous commit.

git-svn-id: https://subversion.assembla.com/svn/emergo@15 fea7a8a5-5848-46ff-b795-545ed800f380