Removed the def-use web construction from the DefUseRules, which was unused.
Implemented the equals/hashCode contract (eclipse generated code) on ValueContainerEdge.

Moved the code that updated the Emergo views in GenerateEmergentInterfaceHandler class into Runnables in order to make it asynchronous.

A new (custom-modified) jw-compiler.jar replaces the old one. The only change is the implementation for the equals/hashCode contract for the Read and Write classes.


git-svn-id: https://subversion.assembla.com/svn/emergo@49 fea7a8a5-5848-46ff-b795-545ed800f380