Optimized DefUseRules by refactoring the on-demand anonymous class instantiations into functors that encapsulate the logic behind the transfer function.
Interprocedural graph depth has been hardcoded to 2. This property must be moved to the project PropertyPage.

The EmergoResultsView, MarkedLinesView are currently broken. Will work on it later.

Some big changes were needed for EmergoView to work as expected now, instead of only working for dependencies in a single compilation unit.

git-svn-id: https://subversion.assembla.com/svn/emergo@51 fea7a8a5-5848-46ff-b795-545ed800f380